# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Circls Platform API
  description: >-
    Circls public + internal API. `/v1/*` is the Firebase-auth internal API used by the admin/partners portals.
    `/api/v1/*` is the aggregator API authenticated with a Circls API key (Bearer ck_…).
  version: 0.1.0
components:
  securitySchemes:
    firebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: Firebase ID token from the partner portal sign-in.
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: ck_… key
      description: Circls API key issued under /v1/tenants/:tenantId/api-keys.
...
paths:
...
  /v1/tenants/{id}/terms/accept:
    post:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        '200':
          description: Default Response
...
  /v1/tenants/{id}:
    get:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        '200':
          description: Default Response
    patch:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        '200':
          description: Default Response
  /v1/tenants/{id}/logo/upload-presign:
    post:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        '200':
          description: Default Response
  /v1/tenants/{id}/logo:
    post:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        '200':
          description: Default Response
    delete:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        '200':
          description: Default Response
  /v1/tenants/{tenantId}/analytics:
    get:
      parameters:
        - schema:
            type: string
          in: path
          name: tenantId
          required: true
      responses:
        '200':
          description: Default Response
  /v1/tenants/{tenantId}/audit-log:
    get:
      parameters:
        - schema:
            type: string
          in: path
          name: tenantId
          required: true
      responses:
        '200':
          description: Default Response
...
servers:
  - url: https://api.circls.app
    description: production
  - url: http://localhost:8080
    description: local
...